+++ gradle.properties
@@ -2,8 +2,11 @@ groupId = dk.mada.mjtar
 version = 0.0.0-SNAPSHOT
 
 org.gradle.caching=true
+org.gradle.configuration-cache=true
 
 dk.mada.style.sonar.projectKey = jskov_mjtar
 dk.mada.style.sonar.organization = jskov-github
 
 dk.mada.style.formatter.include = src/main/java/dk/**/*.java
+dk.mada.style.checkstyle.includes = dk/**/*.java
+dk.mada.style.errorprone.excluded-paths-regexp = .*/org/.*
